storeディレクトリにlocalStorageを移す(コード修正)
